+++ app/src/main/java/com/eelizarraras/workout/core/data/model/entity/WorkoutSetEntity.kt
@@ -4,11 +4,10 @@ import androidx.room.ColumnInfo
 import androidx.room.Entity
 import androidx.room.PrimaryKey
 import com.eelizarraras.workout.core.domine.model.Unit
-import com.eelizarraras.workout.core.domine.model.WorkoutSetModel
 
 @Entity(tableName = "WorkoutSet")
 data class WorkoutSetEntity (
-    @PrimaryKey(autoGenerate = true) val uid: Int,
+    @PrimaryKey(autoGenerate = true) val uid: Long,
     @ColumnInfo(name = "weight") val weight: Double,
     @ColumnInfo(name = "unit") val unit: Unit,
     @ColumnInfo(name = "reps") val reps: Int
