+++ app/src/main/java/com/eelizarraras/workout/core/domine/model/ActivityModel.kt
@@ -1,6 +1,6 @@
 package com.eelizarraras.workout.core.domine.model
 
 data class ActivityModel(
-    val workoutId: Int,
-    val setId: Int
+    val workoutId: Long,
+    val setId: Long
 )
